requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
psycopg2-binary>=2.9.0  # atau mysql-connector-python
openpyxl>=3.1.0
pyswarm>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0